Summarize process errors in a variable
The error type of a process can currently be one of the following:
CONTENT_ERROR
UNKNOWN_ERROR
(in the future we may add more like INFRASTRUCTURE_ERROR, for example)

It can be used when collecting statistics about the Deploy Service or as
a rough indication to the user as to what went wrong. For example, if he
receives a CONTENT_ERROR, he should be able to figure out that the
problem is within the MTA he's trying to deploy.

JIRA: LMCROSSITXSADEPLOY-578