Fix NPE in injectServiceKeysCredentialsInAppEnv
This situation can happen when service broker part of the MTA is a subscriber for a
configuration and it should be restarted in app sub-process.
In app sub-process UpdateAppStep and CreateAppStep steps work with
CloudApplicationExtended which has property "serviceKeysToInject".
But when subscribed service broker are set in context they are stored as
CloudApplication objects which does not have property
"serviceKeysToInject".
The context is set in UpdateSubscribersStep

Jira: LMCROSSITXSADEPLOY-990