Fix Start/End ProcessListeners to not set process finished state earlier.
If there is process with subprocess, there are registered
Start and EndProcessListeners for the parent process and for the subprocess.
In this case the EndProcessListener of the subprocess set the parent
process (correlationId) in finished state earlier, before it is fully
executed. This leads to earlier report for finished state and missing
progress messages in the client. This commit fixes this bug, because
only EndProcessListener of the parent process set the state to finished,
when it is actually fully executed.

JIRA: LMCROSSITXSADEPLOY-1065