Improve 'application not modified' message
The message in it's former state implies that the whole application is
not updated, which is not the case - the app bits may very well be
updated and just a moment later staged, etc.

By specifying that the attributes are not changed and will not be
updated the user gains more insight without being confused by the
following messages.