[Metadata] Ensure services part of new MTA are not marked as discontinued
Having Environment metadata, if a service had no bound applications,
there was no way to determine if that service is part of the MTA as the
service itself does not have environment. With the v3 Metadata, the
service has its own metadata, therefore we can now determine if the
service is part of the MTA or not. This is an enhancement which fixes an
old issue where using --delete-services on orphaned services would
delete them because of the mentioned detection above.