Set csrf headers on the return of the filter chaing
According to this github issue*, the spring security library creates a
new csrf token after user authentication. Moving the copy of that token
value to the response headers to the returning part of the request
filter chain ensures that it will be the correct one.

LMCROSSITXSADEPLOY-1920 CSRF token not properly returned when a session is established

*https://github.com/aditzel/spring-security-csrf-token-interceptor/issues/1#issuecomment-45350315