Add ordering by timestamp for HistoricOperationEvent
Operation events are sorted by their index value by default.
This causes problems in OperationsHelper, which relies on the
operation's last event type being the most up to date.
The issue is that an operation event with an older timestamp
can have a higher index if it is added after the latest event.